requirements.txt
Django==5.1.1
Pillow==10.4.0
sqlparse==0.5.0
asgiref==3.8.1
django-widget-tweaks==1.5.0
django-phonenumber-field==8.0.0
phoneNumbers==8.13.48
twilio==9.2.4
django-otp==1.5.4
